Make the app work when installed
Previously the app would work with `yarn dev` but only display a blank
screen when packaged and installed. This probably stemmed from
differences in configuration between dev and prod environments. This
commit enables Node integration for prod (previously it was enabled only
for dev) and removes an entry in `app/package.json` which duplicated one
in the root `package.json`.